Bugfix: We do no prefilter classes anymore. Take the following example:
List l = new ArrayList();
l.add(secret);
Iterator i = l.iterator();
s = i.next();

In this case, the call to "List.add" must be in the file to make l tainted. However, in the previous version, some method of "Iterator" also had to be in the list for the line "s = i.next()" to be considered. This behaivour is however wrong as "i" itself is tainted, so no matter which method is called, the result will always be tainted.

In "supportsTaintWrappingForClass" we cannot decide that, so we always return "true" there from now on and then defer actually checking what'S going on to "getTaintsForMethod".